Moved SpikeLoader from kwik model to traces waveforms.
